meta-ads-bot/requirements.txt
python-telegram-bot>=20.0
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
groq>=0.5.0
google-generativeai>=0.5.0
edge-tts>=6.1.0
aiohttp>=3.9.0
